Move classes para o pacote de Etiquetas
