+++ src/main/java/br/com/correios/api/etiqueta/EtiquetaCreator.java
@@ -1,0 +1,5 @@
+package br.com.correios.api.etiqueta;
+
+public class EtiquetaCreator {
+
+}
