+++ src/main/java/br/com/correios/api/etiqueta/EtiquetaDosCorreiosExtractor.java
@@ -1,4 +1,4 @@
-package br.com.correios.api.postagem.etiqueta;
+package br.com.correios.api.etiqueta;
 
 import java.util.List;
