--- src/main/java/br/com/correios/api/postagem/etiqueta/EtiquetaCreator.java
@@ -1,5 +1,0 @@
-package br.com.correios.api.postagem.etiqueta;
-
-public class EtiquetaCreator {
-
-}
